+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 Each entry (version) should have a date and one or more of 3 sections: 'Features', 'Bug Fixes', 'Breaking Changes'. Make sure to `git tag` the commit to match the version. Sub / pre-release versions should be hypenated (i.e. 1.0.3-2).
 
+# 1.2.9
+
+## Features
+
+- add in challenge media privacy.
+
+
 # 1.2.8
 
 ## Features